안드로이드 마커 선택 애니메이션에서 줌 레벨이 변경되는 문제 수정

+++ android/src/main/java/com/github/quadflask/react/navermap/RNNaverMapViewManager.java
@@ -14,7 +14,6 @@ import com.facebook.react.uimanager.ViewGroupManager;
 import com.facebook.react.uimanager.annotations.ReactProp;
 import com.naver.maps.geometry.LatLng;
 import com.naver.maps.geometry.LatLngBounds;
-import com.naver.maps.map.CameraUpdate;
 import com.naver.maps.map.NaverMap;
 import com.naver.maps.map.NaverMapOptions;
 import com.naver.maps.map.util.FusedLocationSource;
@@ -291,9 +290,9 @@ public class RNNaverMapViewManager extends ViewGroupManager<RNNaverMapViewContai
                 break;
             case ANIMATE_TO_POSITION:
                 LatLng latLng = toNaverLatLng(args.getMap(0));
-                int zoom = args.getInt(1);
+                Double zoom = args.getDouble(1);
 
-                mapView.scrollToAndZoom(latLng, (float) zoom);
+                mapView.scrollToAndZoom(latLng, zoom.floatValue());
                 break;
             case ANIMATE_TO_COORDINATE:
                 mapView.setCenter(toNaverLatLng(args.getMap(0)));
